PriceList CRUD complete

+++ src/main/java/com/dp3/dao/ProductRepository.java
@@ -1,12 +1,14 @@
 package com.dp3.dao;
 
+import com.dp3.domain.Price;
 import com.dp3.domain.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.math.BigInteger;
+import java.util.Date;
 
 @Repository
-public interface ProductRepository extends JpaRepository<Product, BigInteger> {
+public interface ProductRepository extends JpaRepository<Product, Integer> {
 
 }
